+++ noodles-bam/src/writer/sam_record.rs
@@ -169,12 +169,13 @@ where
 
 fn write_mapping_quality<W>(
     writer: &mut W,
-    mapping_quality: sam::record::MappingQuality,
+    mapping_quality: Option<sam::record::MappingQuality>,
 ) -> io::Result<()>
 where
     W: Write,
 {
-    let mapq = u8::from(mapping_quality);
+    use sam::record::mapping_quality::MISSING;
+    let mapq = mapping_quality.map(u8::from).unwrap_or(MISSING);
     writer.write_u8(mapq)
 }