+++ noodles-cram/src/reader/record.rs
@@ -852,7 +852,9 @@ where
             })
     }
 
-    fn read_mapping_quality(&mut self) -> io::Result<sam::record::MappingQuality> {
+    fn read_mapping_quality(&mut self) -> io::Result<Option<sam::record::MappingQuality>> {
+        use sam::record::mapping_quality::MISSING;
+
         let encoding = self
             .compression_header
             .data_series_encoding_map()
@@ -864,13 +866,19 @@ where
                 )
             })?;
 
-        decode_itf8(
+        let n = decode_itf8(
             encoding,
             &mut self.core_data_reader,
             &mut self.external_data_readers,
         )
-        .and_then(|n| u8::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e)))
-        .map(sam::record::MappingQuality::from)
+        .and_then(|n| u8::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e)))?;
+
+        match n {
+            MISSING => Ok(None),
+            _ => sam::record::MappingQuality::try_from(n)
+                .map(Some)
+                .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e)),
+        }
     }
 
     fn read_unmapped_read(